+++ data/tasks.txt
@@ -2,3 +2,4 @@ T | X | read book
 D |   | return book | 2023-06-06
 E |   | project meeting | Aug 6th 2pm | 4pm
 T | X | join sports club
+T |   | run
